Reworking Highscore system + Adding HighscoreManager + Adjusting HighscoreList + Adjusting GameOverScreen, but not finished it yet - Real high score tracking is still not enabled Moving obsolete files to different directory (keeping them for reference)

+++ Source/Serializable.cpp
@@ -10,7 +10,7 @@
 
 template <> void Serializable::write(std::ostream& stream, std::string input)
 {
-	int length = input.size();
+    unsigned long length = input.size();
 	stream.write((char*)&length, sizeof(length));
 	stream << input;
 }
